Some cleanup, removal of redundant code paths
