+++ Sources/Hyperspace/Request/Decoding/AnyDecodable.swift
@@ -17,6 +17,7 @@ struct AnyDecodable: Decodable {
     }
 }
 
+@available(*, deprecated)
 extension AnyDecodable {
     init(from decoder: Decoder) throws {
         let container = try decoder.singleValueContainer()
